+++ app/pages/Transaction/Tabs/UserTransactionOverviewTab.tsx
@@ -8,7 +8,6 @@ import {
   type CoinDescription,
   useGetCoinList,
 } from "../../../api/hooks/useGetCoinList";
-import {AIP140GasBanner} from "../../../components/AIP140GasBanner";
 import HashButton, {HashType} from "../../../components/HashButton";
 import ContentBox from "../../../components/IndividualPageContent/ContentBox";
 import ContentRow from "../../../components/IndividualPageContent/ContentRow";
@@ -763,7 +762,6 @@ export default function UserTransactionOverviewTab({
 
   return (
     <Box marginBottom={3}>
-      <AIP140GasBanner transaction={transactionData} />
       <ContentBox padding={4}>
         <ContentRow
           title={"Version:"}